:recycle: refactor: 修改所有文件操作为同步

+++ init.ts
@@ -3,11 +3,8 @@ import { getDir } from '@/libs/getDirName.ts'
 import { globalReg } from '@/libs/globalReg.ts'
 import { loadPlugin } from '@/libs/loadPlugin.ts'
 import { getPackage } from '@/libs/loadVersion.ts'
-import { makeSystemLogger } from '@/libs/logger.ts'
-import fs from 'fs/promises'
+import fs from 'fs'
 import path from 'path'
-
-const logger = makeSystemLogger({ pluginName: 'init' })
 
 export default async function () {
   //修改时区
@@ -35,21 +32,13 @@ export default async function () {
 
   // 清空 temp 文件夹
   const tempDir = path.join(global.baseDir, 'temp')
-  await fs
-    .stat(tempDir)
-    .then(async () => {
-      //删除temp文件夹内的所有文件
-      await deleteFolder(tempDir)
-      //创建文件夹
-      await fs.mkdir(tempDir)
-    })
-    .catch(async error => {
-      if (error.toString().includes('no such file or directory')) {
-        //创建文件夹
-        await fs.mkdir(tempDir)
-      } else {
-        logger.WARNING('删除temp文件夹失败')
-        logger.ERROR(error)
-      }
-    })
+
+  if (fs.existsSync(tempDir)) {
+    //删除temp文件夹内的所有文件
+    deleteFolder(tempDir)
+    //创建文件夹
+    fs.mkdirSync(tempDir)
+  } else {
+    fs.mkdirSync(tempDir)
+  }
 }
